hello-flask-redis-project/requirements.txt
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.8
python-dotenv==0.20.0
redis==4.2.2
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
